Make the addition of the LiveJourtnal section faster.
Before I had the Twitter download added as a SCRIPT tag with
the SRC referring to Twitters URL. this meant that the document ready
event did not fire until it had been returned. Changed to use
jQuery's Ajax mechanism, which automatically switches to the
JSONP approach when used cross-site. This means the two sections are
downloaded in parallel.