Changes to better manage the transclusion of LiveJournal entries.
I posted a blog entry boasting of the site being finished and
the very next day LiveJournal changed its format so as to break it!
In the process of trying various fixes -- and (re)discovering that
the Atom feed lacks userpics -- I ended up rewriting the LiveJournal
code in to its own module. The new verison generates much less
nested HTML, since it extracts the infor and formats it from
scratch with a Drupal template, instead of relying on the LJ HTML.