Replace Twitter stream with Timeline widget direct from Twitter.
The API has been changed to require OAuth2 authentication to obtain tweets and
so until I can be bothered to sort that we can just make do with the standard
widget.

Also update stylesheet to allow the Twitter and LiveJoural columns more space
when it is available (as is often the case on modern screens).